requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.4
distlib==0.3.1
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.7
identify==1.4.23
idna==2.10
importlib-metadata==1.7.0
multidict==4.7.6
nodeenv==1.4.0
pathspec==0.8.0
pre-commit==2.6.0
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
smmap==3.0.4
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.27
websockets==8.1
yarl==1.4.2
zipp==3.1.0
